+++ components/AboutComponent/AboutComponent.tsx
