+++ components/SocialMediaIconsComponent/SocialMediaIconsComponent.module.css
@@ -1,0 +1,1 @@
+.SocialMediaIconsComponent {}
