+++ pages/globals.css
