--- src/app/api/hello/route.ts
@@ -1,3 +1,0 @@
-export async function GET(request: Request) {
-  return new Response('Hello, Next.js!')
-}
