--- src/app/layout.tsx
@@ -1,21 +1,0 @@
-import './globals.css'
-
-export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <html lang="en">
-      <body>
-        {/* <p>Paragraph</p> */}
-        {children}
-      </body>
-    </html>
-  )
-}
